More accurately track cross-project observation
Update DefaultLocalComponentRegistry to be dependency-management scoped, and thus is now able to track
and broadcast all cases where one project requests the component metadata of another. This is more accurate
than the existing ResolvedLocalComponentsGraphVisitor, which only considers one project dependent on another
when a project node is present in the final graph. Module and capability conflict resolutions can cause
differences between these two methods of tracking project dependencies, for example if a project was originally
used to influence the graph state but was later removed from the graph.

This seemingly breaks at least one Isolated Projects test. Not sure exactly what that means -- if this exposed
a bug in IP or if there is a bug in this PR. Or if that just means we need to change expectations in the test
now that we are more accurately tracking project accesses.